Make accumulate return uint64_t

+++ flint/Tokenizer.cpp
@@ -15,10 +15,10 @@ namespace std {
     struct hash<flint::StringFragment> {
 
 		typedef flint::StringFragment argument_type;
-		typedef size_t value_type;
+		typedef uint64_t value_type;
 
 		inline value_type operator()(const argument_type &fragment) const {
-			return accumulate(fragment.begin(), fragment.end(), 5381, [](uint64_t curr, char next) {
+			return accumulate(fragment.begin(), fragment.end(), uint64_t(5381), [](uint64_t curr, char next) {
 				return ((curr << 5) + curr) + next;
 			});
 		};
